07-Demo-Application/guestbook-app/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.2
gunicorn==20.1.0
Werkzeug==2.2.2